Refactor SOAP actions in a new .rs file
Fixes #7

Refactor SOAP operations into a new file `src/soap_operations.rs`.

* **Add `src/soap_operations.rs`**:
  - Create functions `send_soap_request` and `process_soap_response` to handle SOAP requests and responses.
* **Modify `src/get_report.rs`**:
  - Remove SOAP request and response handling logic.
  - Call functions in `src/soap_operations.rs` for SOAP operations.
* **Modify `src/main.rs`**:
  - Include the new module `soap_operations`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/zlovtnik/oracle-bip-client/issues/7?shareId=XXXX-XXXX-XXXX-XXXX).
