+++ src/get_report.rs
@@ -1,39 +1,10 @@
-use std::io::Read;
-use base64::engine::general_purpose::STANDARD;
-use base64::Engine;
-use flate2::bufread::GzDecoder;
 use crate::get_xml_by_report::get_xml;
+use crate::soap_operations::{send_soap_request, process_soap_response};
 
 pub fn get_report(url: &str, username: &str, password: &str, report_path: &str, output_format: &str, params: Vec<(&str, String)>)
                   -> Result<Vec<u8>, Box<dyn std::error::Error>> {
-    let client = reqwest::blocking::Client::new();
     let body = get_xml(report_path, output_format, params);
-
-    let res = client
-        .post(url)
-        .header("Content-Type", "text/xml")
-        .header("SOAPAction", "runReport")
-        .header("Authorization", format!("Basic {}", STANDARD.encode(format!("{}:{}", username, password))))
-        .body(body)
-        .send()?;
-
-    let body = res.bytes()?;
-
-    let mut d = GzDecoder::new(&body[..]);
-    let mut decompressed_body = Vec::new();
-    d.read_to_end(&mut decompressed_body)?;
-
-    let decompressed_body = std::str::from_utf8(&decompressed_body)?;
-    println!("{:?}", decompressed_body);
-    let doc = roxmltree::Document::parse(&decompressed_body)?;
-    let node = doc.root_element();
-    let _ns = node.lookup_namespace_uri(None);
-    let mut result = Vec::new();
-    for elem in node.descendants() {
-        if elem.has_tag_name("reportBytes") {
-            let text = elem.text().unwrap();
-            result = STANDARD.decode(text)?;
-        }
-    }
+    let decompressed_body = send_soap_request(url, username, password, body)?;
+    let result = process_soap_response(decompressed_body)?;
     Ok(result)
 }
